Rename directory metric to metrics
